requirements.txt
ruamel.yaml
aiohttp==3.10.10
albumentations==1.4.18
cv_bridge
Cython==3.0.10
cycler
dearpygui==1.11.1
einops==0.8.0
glfw==2.7.0
gdown
h5py
hickle==5.0.3
imageio==2.35.1
joblib==1.4.2
kornia==0.5.10
kiwisolver
loguru==0.7.2
matplotlib==3.9.2
Ninja
numpy==1.26.0
omegaconf==2.3.0
open3d==0.18.0
openai==1.51.2
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.82
opencv_python_headless==4.10.0.84
pandas
Pillow==11.0.0
progressbar33==2.4
psutil
PyAutoGUI==0.9.54
PyOpenGL==3.1.0
pyparsing
pyrender==0.1.45
PySide6==6.8.0.1
pytest==8.2.2
pytinyrenderer==0.0.14
pytorch_lightning==2.4.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3	
ruamel.base==1.0.0
ruptures==1.1.9
scikit_learn==1.5.2
scipy==1.14.1
setuptools==72.1.0
six
tqdm==4.66.5
transformations
trimesh==4.4.7
ultralytics==8.3.31
warp==1.0.4
warp_lang==1.3.1
warp-lang
websocket_client==1.8.0
yacs==0.1.8
PyOpenGL-accelerate==3.1.9
trimesh[easy]
